+++ hopsworksai/resource_cluster_test.go
@@ -1706,7 +1706,8 @@ func TestClusterRead_AWS(t *testing.T) {
 							"clusterConfiguration": {
 								"head": {
 									"instanceType": "node-type-1",
-									"diskSize": 512
+									"diskSize": 512,
+									"nodeId": "head-node-id-1"
 								},
 								"workers": [
 									{
@@ -1752,6 +1753,7 @@ func TestClusterRead_AWS(t *testing.T) {
 				map[string]interface{}{
 					"instance_type": "node-type-1",
 					"disk_size":     512,
+					"node_id":       "head-node-id-1",
 				},
 			},
 			"workers": schema.NewSet(helpers.WorkerSetHash, []interface{}{
@@ -1820,7 +1822,8 @@ func TestClusterRead_AZURE(t *testing.T) {
 							"clusterConfiguration": {
 								"head": {
 									"instanceType": "node-type-1",
-									"diskSize": 512
+									"diskSize": 512,
+									"nodeId": "head-node-id-1"
 								},
 								"workers": [
 									{
@@ -1870,6 +1873,7 @@ func TestClusterRead_AZURE(t *testing.T) {
 				map[string]interface{}{
 					"instance_type": "node-type-1",
 					"disk_size":     512,
+					"node_id":       "head-node-id-1",
 				},
 			},
 			"workers": schema.NewSet(helpers.WorkerSetHash, []interface{}{